+++ Common/Meta.cs
@@ -4,7 +4,7 @@ namespace Dyalect
     {
         public static readonly System.DateTime Epoch = new(2019, 4, 17);
 
-        public const string Version = "0.27.2";
+        public const string Version = "0.28.0";
 
         public const string Product = "Dyalect";
 
@@ -12,7 +12,7 @@ namespace Dyalect
 
         public const string Description = "Dyalect, a dynamic language for .NET Core.";
 
-        public const string Copyright = "Copyright (c) Vasily Voronkov 2019-2021";
+        public const string Copyright = "Copyright (c) Vasily Voronkov 2019-2022";
 
         public const string Trademark = "Vasily Voronkov";
     }