+++ Dyalect/Runtime/Types/DyClass.cs
@@ -29,7 +29,7 @@ namespace Dyalect.Runtime.Types
 
         protected internal override bool HasItem(string name, ExecutionContext ctx) => Fields.HasItem(name, ctx);
 
-        protected internal override DyObject GetTaggedValue() => Fields;
+        protected internal override DyObject Unbox() => Fields;
 
         public override DyObject Clone() => new DyClass(TypeId, Constructor, privateCtor, Fields, DeclaringUnit);
     }