+++ Dyalect/Runtime/Types/DyObject.cs
@@ -36,6 +36,8 @@ namespace Dyalect.Runtime.Types
 
         protected internal virtual DyObject GetTaggedValue() => this;
 
+        protected internal virtual DyObject Unbox() => this;
+        
         public virtual void GetConstructor(ExecutionContext ctx, out string ctor, out bool priv)
         {
             ctor = "";